+++ src/components/DayDrawer/DayDrawer.tsx
@@ -1,10 +1,17 @@
 import dayjs from 'dayjs';
+import localizedFormat from 'dayjs/plugin/localizedFormat';
 import React from 'react';
 import { css } from '@emotion/css';
 import { textUtil } from '@grafana/data';
 import { Badge, Button, Drawer, HorizontalGroup, Icon, LinkButton, useTheme2 } from '@grafana/ui';
 import { CalendarEvent } from '../../types';
 import { CalendarEntry } from '../CalendarEntry';
+
+/**
+ * Day.js Plugins
+ * - https://day.js.org/docs/en/plugin/localized-format
+ */
+dayjs.extend(localizedFormat);
 
 /**
  * Properties
@@ -90,11 +97,9 @@ export const DayDrawer = ({ day, events, event, setEvent, onClose }: Props) => {
      */
     subtitle = `${event.start.format('LLL')}`;
     if (event.end) {
-      if (event.start.startOf('day').isSame(event.end?.startOf('day'))) {
-        subtitle = `${event.start.format('LLL')} - ${event.end.format('LT')}`;
-      }
-
-      subtitle = `${event.start.format('LLL')} - ${event.end.format('LLL')}`;
+      subtitle = `${event.start.format('LLL')} - ${
+        event.start.startOf('day').isSame(event.end?.startOf('day')) ? event.end.format('LT') : event.end.format('LLL')
+      }`;
     }
 
     children = (